♻️ refactor: create a utils file for the limit_mecha
